2608 fix range_window for base_sql backend
